refactor circulating
